reduced duplication line directionscreen
